refactor: extract MeteorUtils from AirdropFlareItem + CustomItemListener
